refactor(path): split path.rs into logical submodules
Split src/utils/path.rs (1600 lines) into 9 smaller modules for
better organization and maintainability.

New structure:
- mod.rs - Main module with re-exports and module documentation
- error.rs - PathError enum and impls
- validate.rs - Path validation functions
- home.rs - Home directory functions (home_dir, home_relative, expand_home)
- format.rs - Path formatting/display functions (shorten_path, abbreviate_path)
- component.rs - Path component accessors (extension, stem, filename, parent, is_hidden)
- util.rs - Path utilities (normalize_separators, join_paths, common_prefix)
- display.rs - PathDisplay struct and impls
- tests.rs - All module tests

All public API items remain accessible through the main module
via re-exports, maintaining backward compatibility.

All 5307 tests pass with no clippy warnings.
