+++ src/utils/path/format.rs
@@ -1,0 +1,155 @@
+use std::path::Path;
+
+use crate::utils::path::util::normalize_separators;
+
+/// Shorten a path to fit within a maximum width
+///
+/// Replaces leading directories with "..." if needed.
+///
+/// # Example
+///
+/// ```rust,ignore
+/// let short = shorten_path("/very/long/path/to/file.txt", 20);
+/// assert_eq!(short, ".../path/to/file.txt");
+/// ```
+pub fn shorten_path(path: impl AsRef<Path>, max_width: usize) -> String {
+    let path = path.as_ref();
+
+    // Convert to display string with normalized separators for consistent handling
+    let full = normalize_separators(&path.display().to_string());
+
+    if full.len() <= max_width {
+        return full;
+    }
+
+    // Try to preserve the filename and as much of the path as possible
+    let components: Vec<&str> = full.split('/').filter(|s| !s.is_empty()).collect();
+
+    if components.is_empty() {
+        return full;
+    }
+
+    // Always keep the filename (safe due to is_empty check above)
+    let filename = match components.last() {
+        Some(f) => f,
+        // This is unreachable due to is_empty check above
+        None => return full,
+    };
+
+    if filename.len() + 4 > max_width {
+        // Even filename doesn't fit, truncate it
+        // Use char_indices for safe UTF-8 truncation
+        let max_bytes = max_width.saturating_sub(3);
+
+        // Find the byte position where we should truncate
+        // We want to include characters that fit within max_bytes
+        let mut truncate_pos = 0;
+        for (byte_start, ch) in filename.char_indices() {
+            let byte_end = byte_start + ch.len_utf8();
+            if byte_end > max_bytes {
+                // This character would exceed the limit
+                break;
+            }
+            // This character fits, include it
+            truncate_pos = byte_end;
+        }
+
+        return format!("...{}", &filename[..truncate_pos]);
+    }
+
+    // Try to include as many trailing components as possible
+    let mut result = String::new();
+    let mut included = 0;
+
+    for (i, component) in components.iter().rev().enumerate() {
+        let needed = if i == 0 {
+            component.len()
+        } else {
+            component.len() + 1 // +1 for separator
+        };
+
+        if result.len() + needed + 4 <= max_width {
+            if i > 0 {
+                result = format!("/{}{}", component, result);
+            } else {
+                result = component.to_string();
+            }
+            included += 1;
+        } else {
+            break;
+        }
+    }
+
+    if included < components.len() {
+        if result.starts_with('/') {
+            format!("...{}", result)
+        } else {
+            format!(".../{}", result)
+        }
+    } else {
+        format!("/{}", result)
+    }
+}
+
+/// Abbreviate path by shortening middle directories to first character
+///
+/// Keeps the last N components fully visible.
+///
+/// # Example
+///
+/// ```rust,ignore
+/// let abbr = abbreviate_path("/Users/john/Documents/Projects/rust/main.rs");
+/// // Returns "/U/j/D/Projects/rust/main.rs"
+/// ```
+pub fn abbreviate_path(path: impl AsRef<Path>) -> String {
+    abbreviate_path_keep(path, 2)
+}
+
+/// Abbreviate path, keeping the last N components fully visible
+pub fn abbreviate_path_keep(path: impl AsRef<Path>, keep_last: usize) -> String {
+    let path = path.as_ref();
+    let path_str = path.display().to_string();
+
+    let starts_with_slash = path_str.starts_with('/');
+    let components: Vec<&str> = path_str.split('/').filter(|s| !s.is_empty()).collect();
+
+    if components.len() <= keep_last {
+        return path_str;
+    }
+
+    let abbrev_count = components.len() - keep_last;
+    let mut result = String::new();
+
+    if starts_with_slash {
+        result.push('/');
+    }
+
+    for (i, component) in components.iter().enumerate() {
+        if i > 0 {
+            result.push('/');
+        }
+
+        if i < abbrev_count {
+            // Abbreviate to first character
+            if let Some(first) = component.chars().next() {
+                result.push(first);
+            }
+        } else {
+            result.push_str(component);
+        }
+    }
+
+    result
+}
+
+/// Get the relative path from a base directory
+pub fn relative_to(path: impl AsRef<Path>, base: impl AsRef<Path>) -> String {
+    let path = path.as_ref();
+    let base = base.as_ref();
+
+    if let Ok(rel) = path.strip_prefix(base) {
+        rel.display().to_string()
+    } else {
+        path.display().to_string()
+    }
+}
