+++ src/utils/path/tests.rs
@@ -1,701 +1,7 @@
-//! Path manipulation utilities
-//!
-//! Provides utilities for displaying and manipulating file paths
-//! in user-friendly formats.
-//!
-//! # Example
-//!
-//! ```rust,ignore
-//! use revue::utils::path::{shorten_path, home_relative, abbreviate_path};
-//!
-//! // Replace home directory with ~
-//! let path = home_relative("/Users/john/Documents/file.txt");
-//! assert_eq!(path, "~/Documents/file.txt");
-//!
-//! // Shorten to fit width
-//! let short = shorten_path("/very/long/path/to/file.txt", 20);
-//! assert_eq!(short, ".../path/to/file.txt");
-//!
-//! // Abbreviate middle directories
-//! let abbr = abbreviate_path("/Users/john/Documents/Projects/rust/src/main.rs");
-//! assert_eq!(abbr, "/U/j/D/P/rust/src/main.rs");
-//! ```
-
-use std::path::{Component, Path, PathBuf};
-
-/// Error types for path validation
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub enum PathError {
-    /// Path contains traversal patterns (..)
-    PathTraversal(String),
-    /// Path contains invalid characters
-    InvalidCharacter(String),
-    /// Path is outside expected bounds
-    OutsideBounds,
-}
-
-impl std::fmt::Display for PathError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            PathError::PathTraversal(p) => write!(f, "Path contains traversal pattern: {}", p),
-            PathError::InvalidCharacter(p) => write!(f, "Path contains invalid characters: {}", p),
-            PathError::OutsideBounds => write!(f, "Path is outside expected bounds"),
-        }
-    }
-}
-
-impl std::error::Error for PathError {}
-
-/// Validate that path doesn't contain traversal patterns
-///
-/// Returns error if the path contains `..` components (ParentDir).
-/// Note: File names containing ".." like "file..txt" are allowed - only actual
-/// parent directory components (..) are rejected.
-pub fn validate_no_traversal(path: &Path) -> Result<(), PathError> {
-    // Check components using the Path API
-    for component in path.components() {
-        if component == Component::ParentDir {
-            return Err(PathError::PathTraversal(path.display().to_string()));
-        }
-    }
-    Ok(())
-}
-
-/// Validate that a path is relative (not absolute)
-///
-/// Returns error if the path is absolute or contains RootDir/Prefix components.
-/// This is a security helper to ensure paths stay within a designated base directory.
-pub fn validate_relative_only(path: &Path) -> Result<(), PathError> {
-    if path.is_absolute() {
-        return Err(PathError::PathTraversal(format!(
-            "Absolute path not allowed: {}",
-            path.display()
-        )));
-    }
-
-    // Also check for RootDir and Prefix components explicitly
-    // This catches edge cases like UNC paths (//server/share) on Unix
-    for component in path.components() {
-        match component {
-            Component::RootDir | Component::Prefix(_) => {
-                return Err(PathError::PathTraversal(format!(
-                    "Root or prefix component not allowed: {}",
-                    path.display()
-                )));
-            }
-            _ => {}
-        }
-    }
-
-    Ok(())
-}
-
-/// Validate that path doesn't contain null bytes
-pub fn validate_characters(path: &Path) -> Result<(), PathError> {
-    let path_str = path.to_string_lossy();
-    if path_str.contains('\0') {
-        return Err(PathError::InvalidCharacter(path.display().to_string()));
-    }
-    Ok(())
-}
-
-/// Find the nearest existing ancestor of a path.
-///
-/// Walks up the directory tree from the given path until an existing directory is found.
-/// Returns the path to that existing ancestor.
-///
-/// # Security
-///
-/// Limits traversal depth to prevent infinite loops with circular symlinks.
-fn find_existing_ancestor(mut path: &Path) -> Option<PathBuf> {
-    const MAX_DEPTH: usize = 100; // Prevent infinite loops
-    let mut depth = 0;
-
-    loop {
-        if path.exists() {
-            return Some(path.to_path_buf());
-        }
-        match path.parent() {
-            Some(parent) if !parent.as_os_str().is_empty() => {
-                path = parent;
-                depth += 1;
-                if depth > MAX_DEPTH {
-                    // Too deep, likely a circular symlink or other issue
-                    return None;
-                }
-            }
-            _ => return None,
-        }
-    }
-}
-
-/// Check if a path is properly within a base directory by prefix matching.
-///
-/// This is a more secure version of `starts_with` that ensures:
-/// 1. The path actually starts with the base prefix
-/// 2. The next character after the prefix is a path separator or end of string
-///    (prevents "/baseevil" from matching "/base")
-fn is_within_base_by_prefix(path: &Path, base: &Path) -> bool {
-    // Use Path::starts_with which handles this correctly
-    // Then verify that if there's more to the path, it starts with a separator
-    if !path.starts_with(base) {
-        return false;
-    }
-
-    // If path equals base exactly, it's within base
-    let path_bytes = path.as_os_str().as_encoded_bytes();
-    let base_bytes = base.as_os_str().as_encoded_bytes();
-
-    if path_bytes.len() == base_bytes.len() {
-        return true;
-    }
-
-    // Check that the next character after base is a path separator
-    // This prevents "/baseevil" from matching "/base"
-    // We check the bytes directly since we know paths are valid UTF-8 or we're checking for ASCII separators
-    if path_bytes.len() > base_bytes.len() {
-        let next_byte = path_bytes[base_bytes.len()];
-        next_byte == b'/' || next_byte == b'\\'
-    } else {
-        false
-    }
-}
-
-/// Validate that a path stays within a base directory
-///
-/// Returns error if the path, when canonicalized, escapes the base directory.
-/// Also checks parent directories for symlinks that might escape the base.
-///
-/// For non-existent paths, walks up the directory tree to find an existing ancestor
-/// and validates that ancestor stays within the base directory.
-///
-/// Returns error if base doesn't exist and path is absolute outside base.
-///
-/// # Security
-///
-/// This function protects against path traversal attacks through multiple mechanisms:
-/// 1. Traversal pattern detection (`..` components)
-/// 2. Symlink escape detection via ancestor canonicalization
-/// 3. Secure prefix matching to prevent prefix bypass attacks
-pub fn validate_within_base(path: &Path, base: &Path) -> Result<(), PathError> {
-    // First validate for traversal patterns
-    validate_no_traversal(path)?;
-
-    // Base must exist for proper validation
-    // If base doesn't exist, we can't reliably validate containment
-    if !base.exists() {
-        // For absolute paths outside a non-existent base, this is suspicious
-        if path.is_absolute() {
-            // Use secure prefix checking to prevent bypass
-            if !is_within_base_by_prefix(path, base) {
-                return Err(PathError::OutsideBounds);
-            }
-        }
-        // For relative paths, we can't validate without base existing
-        // This is a limitation - caller should ensure base exists first
-        return Ok(());
-    }
-
-    // Cache base canonicalization
-    let base_canonical = match base.canonicalize() {
-        Ok(c) => c,
-        Err(_) => {
-            // Base exists but can't canonicalize (permission issues, etc.)
-            // Fall back to secure prefix check
-            if is_within_base_by_prefix(path, base) {
-                return Ok(());
-            }
-            return Err(PathError::OutsideBounds);
-        }
-    };
-
-    // If path exists, validate directly
-    if path.exists() {
-        if let Ok(path_canonical) = path.canonicalize() {
-            if !path_canonical.starts_with(&base_canonical) {
-                return Err(PathError::OutsideBounds);
-            }
-        }
-        return Ok(());
-    }
-
-    // Path doesn't exist - find nearest existing ancestor
-    // If no ancestor exists, we must be more conservative
-    let ancestor = match find_existing_ancestor(path) {
-        Some(a) => a,
-        None => {
-            // No existing ancestor found - path is entirely in non-existent territory
-            // Use prefix checking as a fallback, but also verify no suspicious patterns
-            if is_within_base_by_prefix(path, base) {
-                // Double-check that the relative path from base is safe
-                if let Ok(rel) = path.strip_prefix(base) {
-                    validate_no_traversal(rel)?;
-                    validate_relative_only(rel)?;
-                }
-                return Ok(());
-            }
-            return Err(PathError::OutsideBounds);
-        }
-    };
-
-    // Validate the ancestor is within base
-    if let Ok(ancestor_canonical) = ancestor.canonicalize() {
-        if !ancestor_canonical.starts_with(&base_canonical) {
-            return Err(PathError::OutsideBounds);
-        }
-    }
-
-    // Also validate that the relative path from base doesn't contain suspicious patterns
-    // This catches cases where the path might look legitimate but point outside base
-    if let Ok(rel) = path.strip_prefix(base) {
-        // Double-check no traversal in the relative path
-        validate_no_traversal(rel)?;
-        // Also ensure the relative part is not absolute
-        validate_relative_only(rel)?;
-    }
-
-    Ok(())
-}
-
-/// Get the home directory path
-///
-/// # Security
-///
-/// Reads the HOME or USERPROFILE environment variable. This can be spoofed by
-/// malicious users. Callers should validate that the returned path exists and
-/// is appropriate for their use case.
-///
-/// Returns None if:
-/// - HOME and USERPROFILE are not set
-/// - The path doesn't exist
-/// - The path exists but is not a directory
-pub fn home_dir() -> Option<PathBuf> {
-    let path = std::env::var("HOME")
-        .ok()
-        .map(PathBuf::from)
-        .or_else(|| std::env::var("USERPROFILE").ok().map(PathBuf::from))?;
-
-    // Validate that the path exists and is a directory
-    if path.exists() && path.is_dir() {
-        Some(path)
-    } else {
-        None
-    }
-}
-
-/// Replace home directory with ~ in path display
-///
-/// # Example
-///
-/// ```rust,ignore
-/// let path = home_relative("/Users/john/Documents/file.txt");
-/// // Returns "~/Documents/file.txt" if home is /Users/john
-/// ```
-pub fn home_relative(path: impl AsRef<Path>) -> String {
-    let path = path.as_ref();
-
-    if let Some(home) = home_dir() {
-        if let Ok(stripped) = path.strip_prefix(&home) {
-            return format!("~/{}", stripped.display());
-        }
-    }
-
-    path.display().to_string()
-}
-
-/// Expand ~ to home directory
-///
-/// # Security
-///
-/// Validates that the expanded path doesn't escape the home directory
-/// through path traversal patterns like `~/../../../etc/passwd`.
-///
-/// Returns error if the path contains traversal patterns or would escape home directory.
-///
-/// # Example
-///
-/// ```rust,ignore
-/// // Use with proper error handling:
-/// let path = expand_home(user_input)?;
-///
-/// // For hardcoded safe paths, use unwrap():
-/// let path = expand_home("~/Documents").unwrap();
-/// ```
-pub fn expand_home(path: impl AsRef<Path>) -> Result<PathBuf, PathError> {
-    let path = path.as_ref();
-    let path_str = path.to_string_lossy();
-
-    if let Some(rest) = path_str.strip_prefix("~/") {
-        if let Some(home) = home_dir() {
-            let rest_path = Path::new(rest);
-
-            // Validate the rest doesn't contain traversal patterns
-            validate_no_traversal(rest_path)?;
-
-            // Reject absolute paths after ~/ - they would escape home
-            validate_relative_only(rest_path)?;
-
-            let expanded = home.join(rest_path);
-
-            // Verify the result stays within home directory
-            if home.exists() {
-                if let Ok(home_canonical) = home.canonicalize() {
-                    if expanded.exists() {
-                        if let Ok(expanded_canonical) = expanded.canonicalize() {
-                            if !expanded_canonical.starts_with(&home_canonical) {
-                                return Err(PathError::OutsideBounds);
-                            }
-                        }
-                    }
-                }
-            }
-
-            return Ok(expanded);
-        }
-    } else if path_str == "~" {
-        if let Some(home) = home_dir() {
-            return Ok(home);
-        }
-    }
-
-    Ok(path.to_path_buf())
-}
-
-/// Shorten a path to fit within a maximum width
-///
-/// Replaces leading directories with "..." if needed.
-///
-/// # Example
-///
-/// ```rust,ignore
-/// let short = shorten_path("/very/long/path/to/file.txt", 20);
-/// assert_eq!(short, ".../path/to/file.txt");
-/// ```
-pub fn shorten_path(path: impl AsRef<Path>, max_width: usize) -> String {
-    let path = path.as_ref();
-
-    // Convert to display string with normalized separators for consistent handling
-    let full = normalize_separators(&path.display().to_string());
-
-    if full.len() <= max_width {
-        return full;
-    }
-
-    // Try to preserve the filename and as much of the path as possible
-    let components: Vec<&str> = full.split('/').filter(|s| !s.is_empty()).collect();
-
-    if components.is_empty() {
-        return full;
-    }
-
-    // Always keep the filename (safe due to is_empty check above)
-    let filename = match components.last() {
-        Some(f) => f,
-        // This is unreachable due to is_empty check above
-        None => return full,
-    };
-
-    if filename.len() + 4 > max_width {
-        // Even filename doesn't fit, truncate it
-        // Use char_indices for safe UTF-8 truncation
-        let max_bytes = max_width.saturating_sub(3);
-
-        // Find the byte position where we should truncate
-        // We want to include characters that fit within max_bytes
-        let mut truncate_pos = 0;
-        for (byte_start, ch) in filename.char_indices() {
-            let byte_end = byte_start + ch.len_utf8();
-            if byte_end > max_bytes {
-                // This character would exceed the limit
-                break;
-            }
-            // This character fits, include it
-            truncate_pos = byte_end;
-        }
-
-        return format!("...{}", &filename[..truncate_pos]);
-    }
-
-    // Try to include as many trailing components as possible
-    let mut result = String::new();
-    let mut included = 0;
-
-    for (i, component) in components.iter().rev().enumerate() {
-        let needed = if i == 0 {
-            component.len()
-        } else {
-            component.len() + 1 // +1 for separator
-        };
-
-        if result.len() + needed + 4 <= max_width {
-            if i > 0 {
-                result = format!("/{}{}", component, result);
-            } else {
-                result = component.to_string();
-            }
-            included += 1;
-        } else {
-            break;
-        }
-    }
-
-    if included < components.len() {
-        if result.starts_with('/') {
-            format!("...{}", result)
-        } else {
-            format!(".../{}", result)
-        }
-    } else {
-        format!("/{}", result)
-    }
-}
-
-/// Abbreviate path by shortening middle directories to first character
-///
-/// Keeps the last N components fully visible.
-///
-/// # Example
-///
-/// ```rust,ignore
-/// let abbr = abbreviate_path("/Users/john/Documents/Projects/rust/main.rs");
-/// // Returns "/U/j/D/Projects/rust/main.rs"
-/// ```
-pub fn abbreviate_path(path: impl AsRef<Path>) -> String {
-    abbreviate_path_keep(path, 2)
-}
-
-/// Abbreviate path, keeping the last N components fully visible
-pub fn abbreviate_path_keep(path: impl AsRef<Path>, keep_last: usize) -> String {
-    let path = path.as_ref();
-    let path_str = path.display().to_string();
-
-    let starts_with_slash = path_str.starts_with('/');
-    let components: Vec<&str> = path_str.split('/').filter(|s| !s.is_empty()).collect();
-
-    if components.len() <= keep_last {
-        return path_str;
-    }
-
-    let abbrev_count = components.len() - keep_last;
-    let mut result = String::new();
-
-    if starts_with_slash {
-        result.push('/');
-    }
-
-    for (i, component) in components.iter().enumerate() {
-        if i > 0 {
-            result.push('/');
-        }
-
-        if i < abbrev_count {
-            // Abbreviate to first character
-            if let Some(first) = component.chars().next() {
-                result.push(first);
-            }
-        } else {
-            result.push_str(component);
-        }
-    }
-
-    result
-}
-
-/// Get the relative path from a base directory
-pub fn relative_to(path: impl AsRef<Path>, base: impl AsRef<Path>) -> String {
-    let path = path.as_ref();
-    let base = base.as_ref();
-
-    if let Ok(rel) = path.strip_prefix(base) {
-        rel.display().to_string()
-    } else {
-        path.display().to_string()
-    }
-}
-
-/// Get the file extension (without dot)
-pub fn extension(path: impl AsRef<Path>) -> Option<String> {
-    path.as_ref()
-        .extension()
-        .and_then(|s| s.to_str())
-        .map(|s| s.to_string())
-}
-
-/// Get the filename without extension
-pub fn stem(path: impl AsRef<Path>) -> Option<String> {
-    path.as_ref()
-        .file_stem()
-        .and_then(|s| s.to_str())
-        .map(|s| s.to_string())
-}
-
-/// Get the filename (with extension)
-pub fn filename(path: impl AsRef<Path>) -> Option<String> {
-    path.as_ref()
-        .file_name()
-        .and_then(|s| s.to_str())
-        .map(|s| s.to_string())
-}
-
-/// Get the parent directory
-pub fn parent(path: impl AsRef<Path>) -> Option<String> {
-    path.as_ref().parent().map(|p| p.display().to_string())
-}
-
-/// Check if path is hidden (starts with .)
-pub fn is_hidden(path: impl AsRef<Path>) -> bool {
-    path.as_ref()
-        .file_name()
-        .and_then(|s| s.to_str())
-        .map(|s| s.starts_with('.'))
-        .unwrap_or(false)
-}
-
-/// Normalize path separators (convert \ to /)
-pub fn normalize_separators(path: &str) -> String {
-    path.replace('\\', "/")
-}
-
-/// Join paths with proper separators
-///
-/// # Security
-///
-/// Validates that all path parts don't contain traversal patterns.
-///
-/// Returns error if any path part contains traversal patterns or is absolute.
-///
-/// # Example
-///
-/// ```rust,ignore
-/// // Use with proper error handling:
-/// let path = join_paths(base, &parts)?;
-///
-/// // For hardcoded safe paths, use unwrap():
-/// let path = join_paths(Path::new("/home/user"), &["documents", "file.txt"]).unwrap();
-/// ```
-pub fn join_paths(base: impl AsRef<Path>, parts: &[&str]) -> Result<PathBuf, PathError> {
-    let mut result = base.as_ref().to_path_buf();
-
-    for part in parts {
-        let part_path = Path::new(part);
-
-        // Check for traversal patterns
-        validate_no_traversal(part_path)?;
-
-        // Reject absolute paths - they would escape the base
-        validate_relative_only(part_path)?;
-
-        result = result.join(part_path);
-    }
-
-    Ok(result)
-}
-
-/// Get common prefix of multiple paths
-pub fn common_prefix(paths: &[&Path]) -> Option<PathBuf> {
-    if paths.is_empty() {
-        return None;
-    }
-
-    let first = paths[0];
-    let mut common = PathBuf::new();
-
-    for component in first.components() {
-        let test = common.join(component);
-        if paths.iter().all(|p| p.starts_with(&test)) {
-            common = test;
-        } else {
-            break;
-        }
-    }
-
-    if common.as_os_str().is_empty() {
-        None
-    } else {
-        Some(common)
-    }
-}
-
-/// Path display options
-#[derive(Clone, Debug)]
-pub struct PathDisplay {
-    /// Maximum display width
-    pub max_width: Option<usize>,
-    /// Replace home with ~
-    pub use_tilde: bool,
-    /// Abbreviate middle directories
-    pub abbreviate: bool,
-    /// Number of directories to keep unabbreviated
-    pub keep_dirs: usize,
-}
-
-impl Default for PathDisplay {
-    fn default() -> Self {
-        Self {
-            max_width: None,
-            use_tilde: true,
-            abbreviate: false,
-            keep_dirs: 2,
-        }
-    }
-}
-
-impl PathDisplay {
-    /// Create a new path display config
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    /// Set maximum width
-    pub fn max_width(mut self, width: usize) -> Self {
-        self.max_width = Some(width);
-        self
-    }
-
-    /// Enable/disable tilde for home
-    pub fn tilde(mut self, use_tilde: bool) -> Self {
-        self.use_tilde = use_tilde;
-        self
-    }
-
-    /// Enable/disable abbreviation
-    pub fn abbreviate(mut self, abbrev: bool) -> Self {
-        self.abbreviate = abbrev;
-        self
-    }
-
-    /// Set number of directories to keep
-    pub fn keep(mut self, count: usize) -> Self {
-        self.keep_dirs = count;
-        self
-    }
-
-    /// Format a path according to options
-    pub fn format(&self, path: impl AsRef<Path>) -> String {
-        let mut result = if self.use_tilde {
-            home_relative(path.as_ref())
-        } else {
-            path.as_ref().display().to_string()
-        };
-
-        if self.abbreviate {
-            result = abbreviate_path_keep(&result, self.keep_dirs);
-        }
-
-        if let Some(max) = self.max_width {
-            if result.len() > max {
-                result = shorten_path(&result, max);
-            }
-        }
-
-        result
-    }
-}
-
 #[cfg(test)]
 mod tests {
-    use super::*;
-    use crate::utils::shell::{escape_applescript, escape_powershell};
+    use crate::utils::path::*;
+    use std::path::{Path, PathBuf};
 
     #[test]
     fn test_shorten_path() {
@@ -1106,7 +412,7 @@ mod tests {
         // When path exists, should return the path itself
         if let Some(home) = home_dir() {
             if home.exists() {
-                let result = find_existing_ancestor(&home);
+                let result = crate::utils::path::validate::find_existing_ancestor(&home);
                 assert_eq!(result, Some(home.clone()));
             }
         }
@@ -1118,7 +424,7 @@ mod tests {
         if let Some(home) = home_dir() {
             if home.exists() {
                 let non_existent = home.join("does/not/exist/file.txt");
-                let result = find_existing_ancestor(&non_existent);
+                let result = crate::utils::path::validate::find_existing_ancestor(&non_existent);
                 // Should find home or some existing ancestor
                 assert!(result.is_some(), "Should find an existing ancestor");
                 assert!(
@@ -1132,7 +438,7 @@ mod tests {
     #[test]
     fn test_find_existing_ancestor_empty_path() {
         // Empty path should return None
-        let result = find_existing_ancestor(Path::new(""));
+        let result = crate::utils::path::validate::find_existing_ancestor(Path::new(""));
         assert!(
             result.is_none(),
             "Empty path should have no existing ancestor"
@@ -1444,7 +750,7 @@ mod tests {
         // We can't easily test circular symlinks without filesystem access,
         // but we can verify it doesn't panic or hang on very deep paths
         let deep = Path::new("/a/b/c/d/e/f/g/h/i/j/k/l/m/n/o/p/q/r/s/t/u/v/w/x/y/z");
-        let result = find_existing_ancestor(deep);
+        let result = crate::utils::path::validate::find_existing_ancestor(deep);
         // Should either find an ancestor or return None, but never hang
         // The test passing here means it didn't hang
         let _ = result;
@@ -1585,6 +891,7 @@ mod tests {
     #[test]
     fn test_escape_applescript_unicode() {
         // Test with various Unicode characters
+        use crate::utils::shell::escape_applescript;
         assert_eq!(escape_applescript("Hello 世界"), r#"Hello 世界"#);
         assert_eq!(escape_applescript("Café au lait"), r#"Café au lait"#);
         // Test with emoji (4-byte UTF-8)
@@ -1594,6 +901,7 @@ mod tests {
     #[test]
     fn test_escape_powershell_unicode() {
         // Test with Unicode - should pass through unchanged
+        use crate::utils::shell::escape_powershell;
         assert_eq!(escape_powershell("Hello 世界"), "Hello 世界");
         assert_eq!(escape_powershell("Café"), "Café");
     }
